requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
openai>=1.0.0
pyyaml>=6.0
httpx>=0.25.0
pydantic>=2.0.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
pdf2image>=1.16.0
